audio_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ffmpeg-python==0.2.0
aiofiles==23.2.1

